Using local grunt-cli to build prod war

+++ build-prod-war.sh
@@ -1,16 +1,19 @@
 #!/bin/bash
 
+set -e
+set -x
+
 TMP_DIR=/tmp/bblorganizer
 PROJECT_DIR=$(dirname `which $0`)
 
-rm -rf $TMP_DIR
+[[ -d $TMP_DIR ]] && rm -rf $TMP_DIR
 
 cd $PROJECT_DIR
 git clone ./ $TMP_DIR
 
 cd $TMP_DIR/bblorganizer-web
 npm install
-grunt simpleBuild
+./node_modules/grunt-cli/bin/grunt simpleBuild
 cd $TMP_DIR/bblorganizer-parent
 mvn package -Denv=prod
 mv $TMP_DIR/bblorganizer-web/target/bblorganizer*.war $TMP_DIR/
